feat: refactor card component and styles, replace card import path style: update font family to "Inter" in various components fix: adjust search input and price input styles for better layout
