+++ src/app/account/account.module.css
@@ -41,6 +41,7 @@
   border: none;
   border-radius: 50px;
   padding: 0.7rem 1.7rem;
+  font-family: "Inter", sans-serif;
   font-size: 1.1rem;
   font-weight: 600;
   cursor: pointer;
