+++ src/app/components/aside/aside.module.css
@@ -48,6 +48,7 @@
   }
   
   .userName {
+    font-family: "Inter", sans-serif;
     font-size: 1.25rem;
     font-weight: bold;
     margin-bottom: 0.3rem;
@@ -59,6 +60,7 @@
   }
   
   .memberSince {
+    font-family: "Inter", sans-serif;
     font-size: 0.95rem;
     color: #f7e7e7;
     margin-bottom: 1.5rem;
@@ -74,6 +76,7 @@
     border: 2px solid #fff;
     color: #fff;
     font-size: 1rem;
+    font-family: "Inter", sans-serif;
     cursor: pointer;
     border-radius: 8px;
     padding: 0.6rem 1rem;
@@ -96,6 +99,7 @@
     border: none;
     border-radius: 8px;
     padding: 0.6rem 1rem;
+    font-family: "Inter", sans-serif;
     font-weight: bold;
     cursor: pointer;
     transition: 0.3s;
@@ -118,6 +122,7 @@
     border: 2px solid #fff;
     border-radius: 6px;
     padding: 0.5rem 1rem;
+    font-family: "Inter", sans-serif;
     font-size: 1rem;
     font-weight: 500;
     cursor: pointer;
