--- src/app/components/card/cards.module.css
@@ -1,140 +1,0 @@
-.card {
-  border: 2.5px solid #dbbd9c;
-  border-radius: 18px;
-  background: #fff;
-  box-shadow: 0 2px 10px #dbbd9c22;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  transition: box-shadow 0.2s, transform 0.2s;
-  min-height: 420px;
-}
-
-.card:hover {
-  box-shadow: 0 8px 32px #dbbd9c44;
-  transform: translateY(-6px) scale(1.025);
-}
-
-.imageWrapper {
-  position: relative;
-  width: 100%;
-  height: 180px;
-  background: #faf9f6;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.imageCard {
-  max-width: 100%;
-  max-height: 170px;
-  object-fit: contain;
-  border-radius: 12px;
-}
-
-.categoriaBadge {
-  position: absolute;
-  top: 12px;
-  left: 12px;
-  padding: 0.4rem 1.1rem;
-  border-radius: 20px;
-  color: #fff;
-  font-size: 0.95rem;
-  font-weight: 600;
-  letter-spacing: 1px;
-  box-shadow: 0 2px 8px #dbbd9c33;
-  text-transform: uppercase;
-}
-
-.textCard {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  padding: 1.2rem 1rem 1rem 1rem;
-}
-
-.tituloCard {
-  font-size: 1.2rem;
-  font-weight: 700;
-  color: #f05080;
-  margin-bottom: 0.5rem;
-  min-height: 2.5rem;
-}
-
-.precoMarca {
-  display: flex;
-  align-items: center;
-  gap: 0.7rem;
-  margin-bottom: 0.5rem;
-}
-
-.preco {
-  font-size: 1.15rem;
-  font-weight: 700;
-  color: #5fced4;
-}
-
-.marcaLink {
-  font-size: 1rem;
-  color: #dbbd9c;
-  font-weight: 600;
-  text-decoration: none;
-  border-bottom: 1.5px dotted #dbbd9c;
-  transition: color 0.2s;
-}
-
-.marcaLink:hover {
-  color: #f05080;
-  border-bottom: 1.5px solid #f05080;
-}
-
-.descricaoWrapper {
-  margin-bottom: 1rem;
-  min-height: 48px;
-}
-
-.descricao {
-  color: #555;
-  font-size: 0.98rem;
-  line-height: 1.4;
-}
-
-.cardActions {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.7rem;
-  margin-top: auto;
-}
-
-.botaoCard {
-  background: #fff;
-  color: #f05080;
-  border: 2px solid #dbbd9c;
-  border-radius: 30px;
-  padding: 0.5rem 1.3rem;
-  font-size: 1rem;
-  font-weight: 600;
-  text-decoration: none;
-  transition: background 0.2s, color 0.2s, border 0.2s;
-  cursor: pointer;
-}
-
-.botaoCard:hover {
-  background: #dbbd9c;
-  color: #fff;
-  border-color: #f05080;
-}
-
-.likeButton {
-  background: none;
-  border: none;
-  font-size: 1.5rem;
-  cursor: pointer;
-  margin-left: 0.5rem;
-  transition: transform 0.2s;
-}
-
-.likeButton:hover {
-  transform: scale(1.2);
-}
