+++ src/app/components/cards/index.jsx
@@ -30,9 +30,7 @@ const Card = ({
       <div className={styles.textCard}>
         <h3 className={styles.tituloCard}>{titulo}</h3>
         <div className={styles.precoMarca}>
-          <span className={styles.preco}>
-            R$ {preco?.toFixed(2) || "--"}
-          </span>
+          <span className={styles.preco}>R$ {preco?.toFixed(2) || "--"}</span>
           {marca && (
             <a
               href={linkMarca}
@@ -52,7 +50,10 @@ const Card = ({
           <Link
             href={caminhoArtigo}
             className={styles.botaoCard}
-            style={{ borderColor: corBorda }}
+            style={{
+              borderColor: corBorda,
+              color: corCategoria || "#f05080",
+            }}
           >
             Ver Mais
           </Link>
